refactor: rename marketingHub store and types to intelligenceHub
- Rename useMarketingHubStore to useIntelligenceHubStore
- Rename IMarketingHubSummary → IIntelligenceHubSummary
- Rename IMarketingTotals → IIntelligenceTotals
- Rename IMarketingDateRange → IIntelligenceDateRange
- Update all import references in useLogout.ts
